requirements.txt
bert-extractive-summarizer==0.7.1
blis==0.7.4
catalogue==2.0.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cymem==2.0.5
filelock==3.0.12
idna==2.10
importlib-metadata==4.0.1
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.19.2
packaging==20.9
pandas==1.2.1
pathy==0.5.2
pdfminer.six==20201018
preshed==3.0.5
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.2
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.3.23
srsly==2.4.1
summarizer==0.0.7
thinc==8.0.3
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.8.1
tqdm==4.60.0
transformers==4.5.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wasabi==0.8.2
zipp==3.4.1
